+++ temba/contacts/views.py
@@ -658,7 +658,7 @@ class ContactCRUDL(SmartCRUDL):
     class Export(ModalMixin, OrgPermsMixin, SmartFormView):
 
         form_class = ExportForm
-        submit_button_name = "Export"
+        submit_button_name = _("Export")
         success_url = "@contacts.contact_list"
 
         def derive_params(self):
@@ -895,13 +895,6 @@ class ContactCRUDL(SmartCRUDL):
                         f"{reverse('contacts.contact_update_fields', args=[obj.id])}",
                         on_submit="contactUpdated()",
                     )
-
-            if self.request.user.is_staff:
-                menu.new_group()
-                menu.add_url_post(
-                    _("Service"),
-                    f'{reverse("orgs.org_service")}?organization={obj.org_id}&redirect_url={reverse("contacts.contact_read", args=[obj.uuid])}',
-                )
 
     class Scheduled(OrgObjPermsMixin, SmartReadView):
         """
@@ -1282,7 +1275,7 @@ class ContactCRUDL(SmartCRUDL):
             if new_groups is not None:
                 mods += obj.update_static_groups(new_groups)
 
-            if not self.org.is_anon:
+            if not obj.org.is_anon:
                 urns = []
 
                 for field_key, value in self.form.data.items():