+++ temba/settings_common.py
@@ -466,13 +466,7 @@ GROUP_PERMISSIONS = {
         "orgs.org_spa",
         "orgs.org_surveyor",
     ),
-    "Customer Support": (
-        "campaigns.campaign_read",  # anywhere we allow servicing still needs these
-        "channels.channel_read",
-        "channels.channellog_read",
-        "contacts.contact_read",
-        "flows.flow_editor",
-    ),
+    "Customer Support": (),
     "Granters": ("orgs.org_grant",),
     "Administrators": (
         "airtime.airtimetransfer_list",